agrogreen-ai/requirements.txt
# AgroGreen AI - Requirements
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0

# AI / LangChain
langchain==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5
openai==1.35.7
faiss-cpu==1.8.0
tiktoken==0.7.0

# ML
Pillow==10.3.0
numpy==1.26.4

# Data
pandas==2.2.2

# Frontend
gradio==4.37.1

# Utilities
pydantic==2.7.4
aiofiles==23.2.1
python-multipart==0.0.9
